Updated ENB server
