+++ lib/api/run-server.js
@@ -1,16 +1,15 @@
 var path = require('path'),
-    vow = require('vow'),
-    Server = require('../server/server'),
-    cdir = process.cwd();
+    Server = require('../server'),
+    cwd = process.cwd();
 
 /**
  * Запускает разрабочичий сервер.
  *
  * @param {Object}  [options]
- * @param {String}  [options.mode=development]   Режим сборки.
- * @param {Boolean} [options.cache=true]         Учитывать кэш при запуске таска.
  * @param {String}  [options.dir=process.cwd()]  Корень проекта.
  * @param {Object}  [options.config]             Конфиг сборки. По умолчанию загружается из `.enb/make.js`.
+ * @param {String}  [options.mode=development]   Режим сборки.
+ * @param {Boolean} [options.cache=true]         Учитывать кэш при запуске таска.
  * @param {Number}  [options.port=8080]          Номер порта.
  * @param {String}  [options.host=0.0.0.0]       Имя хоста.
  * @param {String}  [options.socket]             Путь к сокету.
@@ -19,19 +18,19 @@ var path = require('path'),
 module.exports = function (options) {
     options = options || {};
 
-    var server = new Server(),
-        root = path.resolve(options.dir || cdir),
-        opts = {
-            port: options.port || 8080,
-            host: options.host || '0.0.0.0',
+    var root = path.resolve(options.dir || cwd),
+        server = new Server({
+            root: root,
+            port: options.port,
+            host: options.host,
             socket: options.socket,
-            config: options.config,
-            mode: options.mode,
-            cache: options.hasOwnProperty('cache') ? options.cache : true
-        };
-
-    return vow.when(server.init(root, opts))
-        .then(function () {
-            return server.run();
+            enbOptions: {
+                dir: root,
+                config: options.config,
+                mode: options.mode,
+                cache: !options.hasOwnProperty('cache') || options.cache,
+            }
         });
+
+    return server.run();
 };
