+++ lib/server/middleware/enb.js
@@ -1,0 +1,37 @@
+var path = require('path'),
+    cwd = process.cwd(),
+    make = require('../../api/make'),
+    TargetNotFoundError = require('../../errors/target-not-found-error');
+
+module.exports = function (options) {
+    return function (req, res, next) {
+        options = options || {};
+
+        var startTime = new Date(),
+            root = options.root || cwd,
+            pathname = req._parsedUrl.pathname,
+            filename = path.join(root, pathname),
+            target = path.relative(root, filename),
+            targets = target ? [target] : [];
+
+        make(targets, {
+                dir: root,
+                mode: options.mode,
+                cache: true
+            })
+            .then(function () {
+                var endTime = new Date();
+
+                console.log('----- ' + path.normalize(pathname) + ' ' + (endTime - startTime) + 'ms');
+
+                next();
+            })
+            .fail(function (err) {
+                if (err instanceof TargetNotFoundError) {
+                    next();
+                } else {
+                    next(err);
+                }
+            });
+    };
+};
