+++ package.json
@@ -51,10 +51,9 @@
     "glob": "5.0.14",
     "inherit": "2.2.2",
     "lodash": "3.10.1",
-    "mime": "1.2.11",
     "puml-link": "0.0.1",
-    "send": "0.8.3",
-    "serve-static": "1.5.3",
+    "serve-favicon": "2.3.0",
+    "serve-static": "1.10.0",
     "stack-trace": "0.0.9",
     "vow": "0.4.10",
     "vow-fs": "0.3.4",
